+++ pkg/scanner/ports.go
@@ -145,6 +145,7 @@ func initServiceMap() map[int]string {
 		1433:  "mssql",
 		1521:  "oracle",
 		1723:  "pptp",
+		2049:  "nfs",
 		3306:  "mysql",
 		3389:  "ms-wbt-server",
 		4848:  "http",